test(html): abrupt-doctype-public-identifier error
