+++ html/parser/crashtests/doctype/abrupt_doctype_identifier.html
@@ -1,0 +1,2 @@
+<!DOCTYPE html PUBLIC "foo>
+<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01//EN" "foo>
