+++ html/parser/tokenizer.rs
@@ -2320,164 +2320,3 @@ impl Default for HTMLState {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn get_tokenizer_html(
-        input: &'static str,
-    ) -> HTMLTokenizer<impl Iterator<Item = char>> {
-        let stream = InputStreamPreprocessor::new(input.chars());
-        HTMLTokenizer::new(stream)
-    }
-
-    #[test]
-    fn test_simple_tag() {
-        let mut html_tok =
-            get_tokenizer_html(include_str!("crashtests/simple_tag.html"));
-
-        assert_eq!(
-            html_tok.next_token(),
-            Some(HTMLToken::new_start_tag("div".into()))
-        );
-
-        assert_eq!(html_tok.next_token(), Some(HTMLToken::Character('H')));
-        assert_eq!(html_tok.next_token(), Some(HTMLToken::Character('e')));
-        assert_eq!(html_tok.next_token(), Some(HTMLToken::Character('l')));
-        assert_eq!(html_tok.next_token(), Some(HTMLToken::Character('l')));
-        assert_eq!(html_tok.next_token(), Some(HTMLToken::Character('o')));
-        assert_eq!(html_tok.next_token(), Some(HTMLToken::Character(' ')));
-        assert_eq!(html_tok.next_token(), Some(HTMLToken::Character('W')));
-        assert_eq!(html_tok.next_token(), Some(HTMLToken::Character('o')));
-        assert_eq!(html_tok.next_token(), Some(HTMLToken::Character('r')));
-        assert_eq!(html_tok.next_token(), Some(HTMLToken::Character('l')));
-        assert_eq!(html_tok.next_token(), Some(HTMLToken::Character('d')));
-
-        assert_eq!(
-            html_tok.next_token(),
-            Some(HTMLToken::new_end_tag("div".into()))
-        );
-
-        assert_eq!(
-            html_tok.next_token(),
-            Some(HTMLToken::Character('\n'))
-        );
-        assert_eq!(html_tok.next_token(), Some(HTMLToken::EOF));
-    }
-
-    #[test]
-    fn test_simple_tag_attributes() {
-        let mut html_tok = get_tokenizer_html(include_str!(
-            "crashtests/simple_tag_attributes.html"
-        ));
-
-        let attributes: [(String, String); 3] = [
-            (String::from("id"), String::from("un-id")),
-            (
-                String::from("class"),
-                String::from("css-class_1 css-class-2"),
-            ),
-            (String::from("href"), String::from("#")),
-        ];
-
-        assert_eq!(
-            html_tok.next_token(),
-            Some(HTMLToken::StartTag {
-                name: "a".into(),
-                self_closing_flag: false,
-                attributes: attributes.to_vec()
-            })
-        );
-    }
-
-    #[test]
-    fn test_doctype() {
-        let mut html_tok =
-            get_tokenizer_html(include_str!("crashtests/doctype.html"));
-
-        assert_eq!(
-            html_tok.next_token(),
-            Some(HTMLToken::DOCTYPE {
-                name: Some("html".into()),
-                public_identifier: None,
-                system_identifier: None,
-                force_quirks_flag: false
-            })
-        );
-    }
-
-    #[test]
-    fn test_doctype_public() {
-        let mut html_tok = get_tokenizer_html(include_str!(
-            "crashtests/doctype_public.html"
-        ));
-
-        assert_eq!(
-            html_tok.next_token(),
-            Some(HTMLToken::DOCTYPE {
-                name: Some("html".into()),
-                public_identifier: Some(
-                    "-//W3C//DTD HTML 4.01//EN".into()
-                ),
-                system_identifier: Some(
-                    "http://www.w3.org/TR/html4/strict.dtd".into()
-                ),
-                force_quirks_flag: false
-            })
-        );
-
-        html_tok.next_token();
-
-        assert_eq!(
-            html_tok.next_token(),
-            Some(HTMLToken::DOCTYPE {
-                name: Some("math".into()),
-                public_identifier: Some(
-                    "-//W3C//DTD MathML 2.0//EN".into()
-                ),
-                system_identifier: Some(
-                    "http://www.w3.org/Math/DTD/mathml2/mathml2.dtd"
-                        .into()
-                ),
-                force_quirks_flag: false
-            })
-        );
-
-        html_tok.next_token();
-
-        assert_eq!(
-            html_tok.next_token(),
-            Some(HTMLToken::DOCTYPE {
-                name: Some("svg".into()),
-                public_identifier: Some(
-                    "-//W3C//DTD SVG 1.1 Basic//EN".into()
-                ),
-                system_identifier: Some(
-                    "http://www.w3.org/Graphics/SVG/1.1/DTD/svg11-basic.dtd".into()
-                ),
-                force_quirks_flag: false
-            })
-        );
-
-        html_tok.next_token();
-
-        assert_eq!(
-            html_tok.next_token(),
-            Some(HTMLToken::DOCTYPE {
-                name: Some("svg:svg".into()),
-                public_identifier: Some(
-                    "-//W3C//DTD XHTML 1.1 plus MathML 2.0 plus SVG 1.1//EN".into()
-                ),
-                system_identifier: Some(
-                    "http://www.w3.org/2002/04/xhtml-math-svg/xhtml-math-svg.dtd".into()
-                ),
-                force_quirks_flag: false
-            })
-        );
-
-        html_tok.next_token();
-
-        assert_eq!(html_tok.next_token(), Some(HTMLToken::EOF));
-    }
-}
